nosig.1: touch up man page
A few places were using .I instead of .B for command line options.

Don't use .nf/.fi for the list of authors as that implies code to
a lot of people.  This helps with html/markdown generated output.

Along those lines, be consistent in using the exact section names
when referring to them so we can autogenerate links.